repo/backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.6.0
PyMySQL==1.1.1
gunicorn==23.0.0
cryptography==44.0.1
